+++ src/Reservations/Reservations/ChangeLog.md
@@ -21,3 +21,7 @@
 ## Upcoming Release
 * Add billing plan details in getCatalog
 	- new object in the response of get-Catalog
+* Add new Api CalculatePrice
+	- new Api for calculate ReservationOrder price
+* Add new Api Purchase
+	- new Api for Purchase ReservationOrder in powershell